Cleanup CMakeLists, mv TestConfig to its own folder, mv Grammar to implementation file
