+++ CMakeLists.txt
@@ -25,7 +25,7 @@ FIND_PACKAGE( ILCUTIL REQUIRED COMPONENTS ILCSOFT_CMAKE_MODULES )
 INCLUDE( ilcsoft_default_settings )
 INCLUDE( CheckIncludeFileCXX )
 
-FIND_PACKAGE( ROOT 5.27 REQUIRED COMPONENTS MathMore Minuit Minuit2)
+FIND_PACKAGE(ROOT 5.27 REQUIRED)
 LINK_LIBRARIES( ${ROOT_COMPONENT_LIBRARIES} ${ROOT_LIBRARIES} )
 # look for vdt in ROOT INCLUDE, if installed outside of root this will not work
 SET( CMAKE_REQUIRED_INCLUDES ${ROOT_INCLUDE_DIRS} )
@@ -34,42 +34,13 @@ IF( ${ROOT_VDT_FOUND} )
   ADD_DEFINITIONS( "-DCF_USE_VDT" )
 ENDIF()
 
-FIND_PACKAGE( MarlinUtil 1.4 REQUIRED )
-
-FIND_PACKAGE( Boost REQUIRED COMPONENTS regex )
+FIND_PACKAGE(Boost REQUIRED)
 INCLUDE_DIRECTORIES( SYSTEM ${Boost_INCLUDE_DIRS} )
 
-FIND_PACKAGE( Marlin 1.0 REQUIRED ) # minimum required Marlin version
-LINK_LIBRARIES( ${Marlin_LIBRARIES} )
-ADD_DEFINITIONS( ${Marlin_DEFINITIONS} )
-
-FIND_PACKAGE( MarlinTrk REQUIRED )
-INCLUDE_DIRECTORIES(SYSTEM ${MarlinTrk_INCLUDE_DIRS} )
-LINK_LIBRARIES( ${MarlinTrk_LIBRARIES} )
-ADD_DEFINITIONS( ${MarlinTrk_DEFINITIONS} )
-
-FIND_PACKAGE( AIDA REQUIRED )
-LINK_LIBRARIES( ${AIDA_LIBRARIES} )
-ADD_DEFINITIONS( ${AIDA_DEFINITIONS} )
-
-find_package( DD4hep REQUIRED)
-set(CMAKE_MODULE_PATH  ${CMAKE_MODULE_PATH}  ${DD4hep_ROOT}/cmake )
-include( DD4hep )
-
-if(DD4HEP_USE_XERCESC)
-	find_package(XercesC)
-endif()
-include(DD4hep_XML_setup)
-INCLUDE_DIRECTORIES( SYSTEM ${DD4hep_INCLUDE_DIRS} )
-LINK_LIBRARIES( ${DD4hep_LIBRARIES} )
-
-OPTION( MARLINRECO_AIDA "Set to ON to build MarlinReco with AIDA" ON )
-IF( MARLINRECO_AIDA )
-	FIND_PACKAGE( AIDA )
-	IF( NOT AIDA_FOUND )
-		MESSAGE( SEND_ERROR "AIDA not found. Please set MARLINRECO_AIDA to OFF or AIDA_DIR=/path/to/AIDA" )
-	ENDIF()
-ENDIF()
+FIND_PACKAGE(MarlinUtil 1.4 REQUIRED)
+FIND_PACKAGE(Marlin 1.0 REQUIRED)
+FIND_PACKAGE(MarlinTrk REQUIRED)
+FIND_PACKAGE(AIDA REQUIRED)
 
 ### DOCUMENTATION ###########################################################
 
@@ -97,7 +68,7 @@ INSTALL_SHARED_LIBRARY( ${PROJECT_NAME} DESTINATION lib )
 
 
 # optional package
-FOREACH( pkg Marlin MarlinUtil AIDA ROOT )
+FOREACH( pkg Marlin MarlinTrk MarlinUtil AIDA)
 	IF( ${pkg}_FOUND )
 		INCLUDE_DIRECTORIES(SYSTEM ${${pkg}_INCLUDE_DIRS})
 		TARGET_LINK_LIBRARIES(${PROJECT_NAME} PRIVATE ${${pkg}_LIBRARIES})
@@ -112,7 +83,7 @@ ENDFOREACH()
 # display some variables and write them to cache
 DISPLAY_STD_VARIABLES()
 
-ADD_EXECUTABLE(TestConfig src/TestConfig.cpp)
+ADD_EXECUTABLE(TestConfig source/test/TestConfig.cpp)
 TARGET_INCLUDE_DIRECTORIES(TestConfig SYSTEM PRIVATE ${Boost_INCLUDE_DIRS})
 
 ### FORMATTING ##############################################################
